requirements.txt
kivy==1.11.1
pdfminer3k
xlrd
python_docx
docx
beautifulsoup4
matplotlib==2.2.2
numpy
